Add useAddedTags hook
